+++ actix/CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Add `Mocker::mock_one()`.
 - Derive `PartialEq` and `Eq` for `MailboxError`.
 - Minimum supported Rust version (MSRV) is now 1.65.